requirements.txt
msal
requests
python-dotenv
pymupdf
python-docx
python-pptx
openpyxl
pinecone[grpc]
sentence-transformers
mcp
pydantic-settings
aiohttp
Pillow
easyocr
azure-ai-vision-imageanalysis
azure-ai-formrecognizer
olefile
markdown
